[WIP] Allow logical name along with UUID in port creation
Currently a port can be created via a POST /v1/ports, with 'node_uuid' and
'address' specified in the body. Now that we have logical names, we should
allow the port creation to accept the name of the node instead of the UUID of
the node. E.g. In request body pass "node_name" instead of "node_uuid"

Change-Id: I531001483c1b592a63ba5a9283f8e37ca07fdc69